Split SimplifiedConsoleOutputDeviceBase into partial class files
Fixes #10206

The 563-line SimplifiedConsoleOutputDeviceBase.cs is split into four
focused partial class files, each under 300 lines:

- SimplifiedConsoleOutputDeviceBase.cs (149): fields, constructors,
  extension identity, initialization, abstract console members and
  process role handling.
- SimplifiedConsoleOutputDeviceBase.Session.cs (151): session lifecycle
  display (banner, before/after session, session start/finish).
- SimplifiedConsoleOutputDeviceBase.TestEvents.cs (231): data
  consumption and output device message display.
- SimplifiedConsoleOutputDeviceBase.SlowTests.cs (71): background
  slow-test detection and reporting.

This is a pure code move: no behavior, API or member accessibility
changes.

Copilot-Session: 50b83f93-7ffb-45db-933d-fa9d34c520d5
